Easier way to get started reading/writing proc. memoru
This is a tool that is designed for newcomers to use to help make things easier to understand/do. :P

How to use it  :
Hit the enter key, any key after enter is fine

Don't worry if you think you will need to have lightning quick reflexes, this is designed to wait for you until you are ready.

The variable "x", is what you are looking for.

Yes, Ik, "using namespace std;" is a bad thing to get into but to be fair I am a lazy mother fucker.